WIP ambulance request into emergency referrals
