+++ mwana/apps/smgl/keyword_handlers/__init__.py
@@ -1,10 +1,8 @@
-from .ambulance_workflow import (ambulance_request, ambulance_response,
-    ambulance_outcome)
 from .birth import birth_registration
 from .contact import leave, make_active, out
 from .death import death_registration
 from .postpartum import postpartum_visit
 from .pregnancy import pregnant_registration, follow_up, motherid_lookup
-from .referrals import refer, referral_outcome
+from .referrals import refer, referral_outcome, emergency_response
 from .syphilis import syphilis
 from .told import told
