+++ mwana/apps/smgl/keyword_handlers/referrals.py
@@ -1,6 +1,7 @@
+import logging
+
 from rapidsms.messages import OutgoingMessage
-from mwana.apps.smgl.app import FACILITY_NOT_RECOGNIZED
-from mwana.apps.smgl.models import Referral
+from mwana.apps.smgl.models import Referral, AmbulanceRequest, AmbulanceResponse
 from mwana.apps.smgl.utils import get_location, to_time, get_session_message
 from mwana.apps.smgl import const
 from mwana.apps.contactsplus.models import ContactType
@@ -8,6 +9,19 @@ from rapidsms.models import Contact
 from datetime import datetime
 from mwana.apps.smgl.decorators import registration_required, is_active
 from django.template.defaultfilters import yesno
+
+from mwana.apps.smgl.app import (get_value_from_form, send_msg, ER_TO_DRIVER,
+    ER_TO_TRIAGE_NURSE, ER_STATUS_UPDATE,
+    INITIAL_AMBULANCE_RESPONSE, _get_allowed_ambulance_workflow_contact,
+    NOT_REGISTERED_TO_CONFIRM_ER, ER_CONFIRM_SESS_NOT_FOUND, ER_TO_CLINIC_WORKER,
+    AMB_OUTCOME_ORIGINATING_LOCATION_INFO, AMB_OUTCOME_FILED, FACILITY_NOT_RECOGNIZED,
+    AMB_RESPONSE_ORIGINATING_LOCATION_INFO, AMB_RESPONSE_NOT_AVAILABLE)
+
+logger = logging.getLogger(__name__)
+# In RapidSMS, message translation is done in OutgoingMessage, so no need
+# to attempt the real translation here.  Use _ so that makemessages finds
+# our text.
+_ = lambda s: s
 
 
 @registration_required
@@ -18,8 +32,8 @@ def refer(session, xform, router):
     assert session.connection.contact.location is not None, \
         "Contact must have a location to refer"
     get_session_message(session)
-
-    name = session.connection.contact.name
+    contact = session.connection.contact
+    name = contact.name
 
     mother_id = xform.xpath("form/unique_id")
     facility_id = xform.xpath("form/facility")
@@ -37,7 +51,21 @@ def refer(session, xform, router):
         if reasons:
             for r in reasons.split(" "):
                 referral.set_reason(r)
-        referral.status = xform.xpath("form/status")
+        status = xform.xpath("form/status")
+        referral.status = status
+        if referral.status == 'em':
+            #DO SOMETHING WITH AMBULANCES
+            session.template_vars.update({"sender_phone_number": session.connection.identity})
+            amb = AmbulanceRequest()
+            amb.session = session
+            session.template_vars.update({"from_location": str(contact.location.name)})
+            amb.set_mother(mother_id)
+            amb.save()
+            referral.amb_req = amb
+            #Respond that we're on it.
+            send_msg(session.connection, INITIAL_AMBULANCE_RESPONSE, router, **session.template_vars)
+            _broadcast_to_ER_users(amb, session, xform, router=router)
+
         try:
             referral.time = to_time(xform.xpath("form/time"))
         except ValueError, e:
@@ -67,7 +95,8 @@ def refer(session, xform, router):
 @is_active
 def referral_outcome(session, xform, router):
     get_session_message(session)
-    name = session.connection.contact.name if session.connection.contact else ""
+    contact = session.connection.contact
+    name = contact.name
     mother_id = xform.xpath("form/unique_id")
 
     refs = Referral.objects.filter(mother_uid=mother_id.lower()).order_by('-date')
@@ -96,27 +125,115 @@ def referral_outcome(session, xform, router):
         ref.mode_of_delivery = xform.xpath("form/mode_of_delivery")
 
     ref.save()
-    router.outgoing(OutgoingMessage(session.connection,
-                                    const.REFERRAL_OUTCOME_RESPONSE % \
-                                        {'name': name, "unique_id": mother_id}))
-    get_session_message(session, direction='O')
 
-    # also notify folks at the referring facility about the outcome
-    for c in _get_people_to_notify_outcome(ref):
-        if c.default_connection:
-            if ref.mother_showed:
-                router.outgoing(OutgoingMessage(c.default_connection,
-                                                const.REFERRAL_OUTCOME_NOTIFICATION % \
-                                                    {"unique_id": ref.mother_uid,
-                                                     "date": ref.date.date(),
-                                                     "mother_outcome": ref.get_mother_outcome_display(),
-                                                     "baby_outcome": ref.get_baby_outcome_display(),
-                                                     "delivery_mode": ref.get_mode_of_delivery_display()}))
+    req = ref.ambulancerequest_set.all()
+    if req:
+        req = req[0]
+        if req.ambulanceresponse_set.all()[0].response != 'na':
+            session.template_vars.update({"contact_type": contact.types.all()[0],
+                                          "name": contact.name})
+
+            if ref.ambulancerequest_set.all().ambulance:
+                _broadcast_to_ER_users(req, session, xform,
+                                       message=_(AMB_OUTCOME_FILED), router=router)
+            send_msg(req.contact.default_connection, AMB_OUTCOME_ORIGINATING_LOCATION_INFO, router, **session.template_vars)
+            return True
+    else:
+        router.outgoing(OutgoingMessage(session.connection,
+                                        const.REFERRAL_OUTCOME_RESPONSE % \
+                                            {'name': name, "unique_id": mother_id}))
+        get_session_message(session, direction='O')
+
+        # also notify folks at the referring facility about the outcome
+        for c in _get_people_to_notify_outcome(ref):
+            if c.default_connection:
+                if ref.mother_showed:
+                    router.outgoing(OutgoingMessage(c.default_connection,
+                                                    const.REFERRAL_OUTCOME_NOTIFICATION % \
+                                                        {"unique_id": ref.mother_uid,
+                                                         "date": ref.date.date(),
+                                                         "mother_outcome": ref.get_mother_outcome_display(),
+                                                         "baby_outcome": ref.get_baby_outcome_display(),
+                                                         "delivery_mode": ref.get_mode_of_delivery_display()}))
+                else:
+                    router.outgoing(OutgoingMessage(c.default_connection,
+                                                    const.REFERRAL_OUTCOME_NOTIFICATION_NOSHOW % \
+                                                        {"unique_id": ref.mother_uid,
+                                                         "date": ref.date.date()}))
+    return True
+
+
+@registration_required
+@is_active
+def emergency_response(session, xform, router):
+    """
+    This handler deals with a status update from an ER Driver or Triage Nurse
+    about a specific ambulance
+
+    i.e. Ambulance on the way/delayed/not available
+    """
+    logger.debug('POST PROCESSING FOR RESP KEYWORD')
+    connection = session.connection
+    contact = _get_allowed_ambulance_workflow_contact(session)
+    if not contact:
+        send_msg(connection, NOT_REGISTERED_TO_CONFIRM_ER, router, **session.template_vars)
+        return True
+
+    ambulance_response = AmbulanceResponse()
+    ambulance_response.responder = contact
+    ambulance_response.session = session
+
+    unique_id = get_value_from_form('unique_id', xform)
+    ambulance_response.mother_uid = unique_id
+    session.template_vars.update({'unique_id': unique_id})
+    #try match uid to mother
+    ambulance_response.set_mother(unique_id)
+    ambulance_response.mother_uid = unique_id
+
+    status = get_value_from_form('status', xform).lower()
+    ambulance_response.response = status
+    session.template_vars.update({"status": status.upper(),
+                                   "response": status.upper()})
+
+    #we might be dealing with a mother that has gone through ER multiple times
+    ambulance_requests = AmbulanceRequest.objects.filter(mother_uid=unique_id)\
+                .exclude(received_response=True)\
+                .order_by('-requested_on')
+    if not ambulance_requests.count():
+        #session doesn't exist or it has already been confirmed
+        send_msg(connection, ER_CONFIRM_SESS_NOT_FOUND, router, **session.template_vars)
+        return True
+
+    #take the latest one in case this mother has been ER'd a bunch
+    ambulance_response.ambulance_request = ambulance_request = ambulance_requests[0]
+    ambulance_request.received_response = True
+
+    confirm_contact_type = contact.types.all()[0]
+    session.template_vars.update({"confirm_type": confirm_contact_type,
+                                  "name": contact.name})
+
+    _broadcast_to_ER_users(ambulance_request, session, xform,
+        message=_(ER_STATUS_UPDATE), router=router)
+
+    ambulance_request.save()
+    ambulance_response.save()
+    referrer_contact = ambulance_request.contact
+    send_msg(referrer_contact.default_connection,
+             AMB_RESPONSE_ORIGINATING_LOCATION_INFO,
+             router, **session.template_vars)
+    if status == 'na':
+        session.template_vars.update({"sender_phone_number": ambulance_request.contact.default_connection.identity,
+                                      "from_location": str(ambulance_request.contact.location.name)})
+        for su in _pick_superusers(session, xform, ambulance_request.receiving_facility):
+            send_msg(su.default_connection, AMB_RESPONSE_NOT_AVAILABLE, router, **session.template_vars)
+    else:
+        clinic_recip = _pick_clinic_recip(session, xform, ambulance_request.receiving_facility)
+        if clinic_recip:
+            ambulance_request.receiving_facility_recipient = clinic_recip
+            if clinic_recip.default_connection:
+                send_msg(clinic_recip.default_connection, ER_TO_CLINIC_WORKER, router, **session.template_vars)
             else:
-                router.outgoing(OutgoingMessage(c.default_connection,
-                                                const.REFERRAL_OUTCOME_NOTIFICATION_NOSHOW % \
-                                                    {"unique_id": ref.mother_uid,
-                                                     "date": ref.date.date()}))
+                logger.error('No Receiving Clinic Worker found (or missing connection for Ambulance Session: %s, XForm Session: %s, XForm: %s' % (ambulance_response, session, xform))
     return True
 
 
@@ -146,3 +263,69 @@ def _get_people_to_notify_outcome(referral):
        referral.get_connection().contact not in facility_contacts:
         facility_contacts.append(referral.get_connection().contact)
     return facility_contacts
+
+
+def _pick_er_driver(session, xform, receiving_facility):
+    ad_type = ContactType.objects.get(slug__iexact='am')
+    ads = Contact.objects.filter(types=ad_type, location=receiving_facility)
+    if ads.count():
+        return ads[0]
+    else:
+        raise Exception('No Ambulance Driver type found!')
+
+
+def _pick_er_triage_nurse(session, xform, receiving_facility):
+    tn_type = ContactType.objects.get(slug__iexact='tn')
+    tns = Contact.objects.filter(types=tn_type, location=receiving_facility)
+    if tns.count():
+        return tns[0]
+    else:
+        raise Exception('No Triage Nurse type found!')
+
+
+def _pick_clinic_recip(session, xform, receiving_facility):
+    cw_type = ContactType.objects.get(slug__iexact='worker')
+    cws = Contact.objects.filter(types=cw_type, location=receiving_facility)
+    if cws.count():
+        return cws[0]
+    else:
+        logger.error('No clinic worker found!')
+
+
+def _pick_superusers(session, xform, receiving_facility):
+    superusers = Contact.objects.filter(is_super_user=True,
+                                        location=receiving_facility)
+    if superusers.count():
+        return superusers
+    else:
+        raise Exception('No Ambulance Driver type found!')
+
+
+def _broadcast_to_ER_users(ambulance_session, session, xform, router, message=None):
+    """
+    Broadcasts a message to the Emergency Response users.  If message is not
+    specified, will send the default initial ER message to each respondent.
+    """
+    receiving_facility = ambulance_session.receiving_facility
+    ambulance_driver = _pick_er_driver(session, xform, receiving_facility)
+    ambulance_session.ambulance_driver = ambulance_driver
+    if ambulance_driver.default_connection:
+        if message:
+            send_msg(ambulance_driver.default_connection, message, router, **session.template_vars)
+        else:
+            send_msg(ambulance_driver.default_connection, ER_TO_DRIVER, router, **session.template_vars)
+    else:
+        logger.error('No Ambulance Driver found (or missing connection) for Ambulance Session: %s, XForm Session: %s, XForm: %s' % (ambulance_session, session, xform))
+
+    tn = _pick_er_triage_nurse(session, xform, receiving_facility)
+    ambulance_session.triage_nurse = tn
+
+    if tn.default_connection:
+        if message:
+            send_msg(tn.default_connection, message, router, **session.template_vars)
+        else:
+            send_msg(tn.default_connection, ER_TO_TRIAGE_NURSE, router, **session.template_vars)
+    else:
+        logger.error('No Triage Nurse found (or missing connection) for Ambulance Session: %s, XForm Session: %s, XForm: %s' % (ambulance_session, session, xform))
+
+    ambulance_session.save()
