+++ mwana/apps/smgl/tests/__init__.py
@@ -1,4 +1,3 @@
-from .test_ambulance_workflow import *
 from .test_birth_reg import *
 from .test_bug_reports import *
 from .test_death_reg import *
