--- mwana/xforms/amb.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<h:html xmlns:h="http://www.w3.org/1999/xhtml" xmlns:orx="http://openrosa.org/jr/xforms" xmlns="http://www.w3.org/2002/xforms" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:jr="http://openrosa.org/javarosa">
-	<h:head>
-		<h:title>AMB</h:title>
-		<model>
-			<instance>
-				<data xmlns:jrm="http://dev.commcarehq.org/jr/xforms" xmlns="http://openrosa.org/formdesigner/A652487B-4E6F-4802-A56F-F68BC7207A47" uiVersion="1" version="1" name="AMB">
-					<unique_id />
-					<danger_signs />
-				</data>
-			</instance>
-			<bind nodeset="/data/unique_id" type="xsd:string" />
-			<bind nodeset="/data/danger_signs" type="xsd:string" />
-			<itext>
-				<translation lang="en" default="">
-					<text id="unique_id-label">
-						<value>Mother's Unique ID</value>
-					</text>
-					<text id="danger_signs-label">
-						<value>Danger Signs</value>
-					</text>
-				</translation>
-			</itext>
-		</model>
-	</h:head>
-	<h:body>
-		<input ref="/data/unique_id">
-			<label ref="jr:itext('unique_id-label')" />
-		</input>
-		<input ref="/data/danger_signs">
-			<label ref="jr:itext('danger_signs-label')" />
-		</input>
-	</h:body>
-</h:html>
