--- mwana/xforms/amb_outcome.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<h:html xmlns:h="http://www.w3.org/1999/xhtml" xmlns:orx="http://openrosa.org/jr/xforms" xmlns="http://www.w3.org/2002/xforms" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:jr="http://openrosa.org/javarosa">
-	<h:head>
-		<h:title>Ambulance Outcome</h:title>
-		<model>
-			<instance>
-				<data xmlns:jrm="http://dev.commcarehq.org/jr/xforms" xmlns="http://openrosa.org/formdesigner/F6901DC3-3BBD-4ABE-A983-E26FC6FFAE6E" uiVersion="1" version="1" name="Ambulance Outcome">
-					<unique_id />
-					<outcome />
-					<no_amb />
-				</data>
-			</instance>
-			<bind nodeset="/data/unique_id" type="xsd:string" />
-			<bind nodeset="/data/outcome" type="xsd:string" />
-			<bind nodeset="/data/no_amb" type="xsd:string" />
-			<itext>
-				<translation lang="en" default="">
-					<text id="unique_id-label">
-						<value>Mother's Unique ID</value>
-					</text>
-					<text id="outcome-label">
-						<value>Outcome</value>
-					</text>
-					<text id="no_amb-label">
-						<value>No Ambulance?</value>
-					</text>
-				</translation>
-			</itext>
-		</model>
-	</h:head>
-	<h:body>
-		<input ref="/data/unique_id">
-			<label ref="jr:itext('unique_id-label')" />
-		</input>
-		<input ref="/data/outcome">
-			<label ref="jr:itext('outcome-label')" />
-		</input>
-		<input ref="/data/no_amb">
-			<label ref="jr:itext('no_amb-label')" />
-		</input>
-	</h:body>
-</h:html>
